libs/flowrate: Move into internal
