+++ internal/flowrate/util.go
